5/22 - Creation of Glossary, revision of ActivateTypes
The Glossary section of the Layout file has been removed from all versions and converted into its own .md file. It will live in the Attributes folder alongside the other files, unless future circumstances require it be moved.

The glossary is obviously incomplete, and will have definitions added to it as time goes on. A good way to keep track of what things should and shouldn't be called.

Revised the language on ActivateTypes.